refacto(bull): Split code & upgrade
