+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
 Copyright (c) 2020 Igor Golovin
+Copyright (c) 2023 Vincent Le Quec (Venatum)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
