+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bullboard",
-  "version": "2.0.3",
+  "version": "5.0.0",
   "main": "src/index.js",
   "license": "MIT",
   "scripts": {
@@ -8,11 +8,11 @@
     "dev": "nodemon -r esm ."
   },
   "dependencies": {
-    "@bull-board/api": "^3.2.6",
-    "@bull-board/express": "^3.2.6",
+    "@bull-board/api": "^5.0.0",
+    "@bull-board/express": "^5.0.0",
     "body-parser": "^1.20.2",
-    "bull": "^3.13.0",
-    "bullmq": "^1.8.4",
+    "bull": "^4.10.4",
+    "bullmq": "^3.10.1",
     "connect-ensure-login": "^0.1.1",
     "dotenv": "^16.0.3",
     "esm": "^3.2.25",
