+++ src/index.js
@@ -4,45 +4,9 @@ import session from 'express-session';
 import bodyParser from 'body-parser';
 import { ensureLoggedIn } from 'connect-ensure-login';
 
-import {redisConfig, client} from "./redis";
 import {config} from "./config";
 import {authRouter} from './login';
-
-
-const {createBullBoard} = require('@bull-board/api');
-const {BullAdapter} = require('@bull-board/api/bullAdapter');
-const {BullMQAdapter} = require('@bull-board/api//bullMQAdapter');
-const {ExpressAdapter} = require('@bull-board/express');
-const Queue = require('bull');
-const bullmq = require('bullmq');
-
-const serverAdapter = new ExpressAdapter();
-const {setQueues} = createBullBoard({queues: [], serverAdapter});
-const router = serverAdapter.getRouter();
-
-async function main() {
-	// await client.connect();
-
-	client.KEYS(`${config.BULL_PREFIX}:*`, (err, keys) => {
-		const uniqKeys = new Set(keys.map(key => key.replace(/^.+?:(.+?):.+?$/, '$1')));
-		const queueList = Array.from(uniqKeys).sort().map(
-			(item) => {
-				if (config.BULL_VERSION === 'BULLMQ') {
-					return new BullMQAdapter(new bullmq.Queue(item, {connection: redisConfig.redis}));
-				}
-
-				return new BullAdapter(new Queue(item, redisConfig));
-			}
-		);
-
-		setQueues(queueList);
-		console.log('done!')
-	});
-}
-
-main();
-
-// ------------------------------------
+import {router} from "./bull";
 
 const app = express();
